Fix incorrect application of amplitude for Ring and Morph modulation.
This has been broken for a while (5c03380247f7), it's kind of
surprising that no one has discovered it before. The amplitude should
not be applied to Ring and Morph modulation because they apply it
themselves in their respective "apply" functions. Not only is it
applied twice, but also incorrectly.

Fixed by simply skipping amplitude for these modulations and letting
it happen in the mentioned "apply" functions.

Signed-off-by: Kristian Amlie <kristian@amlie.name>